+++ src/components/MessageTo/CreateButton/CreateButton.module.css
@@ -1,0 +1,37 @@
+@import url("../../../styles/color.css");
+
+.CreateButton {
+  display: flex;
+  width: 72rem;
+  padding: 1.4rem 2.4rem;
+  margin: 2.4rem;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+
+  border-radius: 1.2rem;
+  background: var(--Purple-60);
+
+  color: var(--Surface);
+  text-align: center;
+
+  /* Font/18 Bold */
+  font-family: Pretendard;
+  font-size: 1.8rem;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 2.8rem; /* 155.556% */
+  letter-spacing: -0.018rem;
+}
+
+.CreateButton:disabled {
+  background: var(--Gray-30);
+}
+
+@media (min-width: 357px) and (max-width: 767px) {
+  .CreateButton {
+    margin-left: 0;
+    margin-right: 0;
+    width: 32rem;
+  }
+}
