+++ src/components/MessageTo/MessageTo.module.css
@@ -47,30 +47,6 @@
   letter-spacing: -0.016rem;
 }
 
-.CreateButton {
-  display: flex;
-  width: 72rem;
-  padding: 1.4rem 2.4rem;
-  margin: 2.4rem;
-  justify-content: center;
-  align-items: center;
-  gap: 1rem;
-
-  border-radius: 1.2rem;
-  background: var(--Purple-60);
-
-  color: var(--Surface);
-  text-align: center;
-
-  /* Font/18 Bold */
-  font-family: Pretendard;
-  font-size: 1.8rem;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 2.8rem; /* 155.556% */
-  letter-spacing: -0.018rem;
-}
-
 @media (min-width: 768px) and (max-width: 1199px) {
   .PostPageBody {
     margin-top: 4.9rem;
@@ -92,11 +68,5 @@
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
-  }
-
-  .CreateButton {
-    margin-left: 0;
-    margin-right: 0;
-    width: 32rem;
   }
 }
